# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Delivery",
        "title": "Food-Delivery API",
        "contact": {},
        "version": "1.0"
    },
    "host": "165.232.68.67:8081",
    "basePath": "/",
    "paths": {
        "/customer": {
            "post": {
                "description": "Create customer with provided data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer"
                ],
                "parameters": [
                    {
                        "description": "Customer data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Customer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    }
                }
            }
        },
        "/customer/orders/{id}": {
            "get": {
                "description": "Get Orders by customer ID and order status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/handler.ordersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.dataResponse"
                        }
                    }
                }
            }
        },
        "/customer/{id}": {
            "get": {
                "description": "Get customer data by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/handler.customerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.dataResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update customer data by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Customer data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CustomerUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/handler.customerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.dataResponse"
                        }
                    }
                }
            }
        },
        "/order": {
            "post": {
                "description": "Create new order",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Create new order",
                "parameters": [
                    {
                        "description": "Order info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/handler.dataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    }
                }
            }
        },
        "/order/{id}": {
            "get": {
                "description": "Get Order by order ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Get Order by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/handler.orderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Customer": {
            "type": "object",
            "required": [
                "address",
                "fullName",
                "id",
                "phoneNumber"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phoneNumber": {
                    "type": "integer"
                }
            }
        },
        "domain.CustomerInfo": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "birthday": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                }
            }
        },
        "domain.CustomerUpdate": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "birthday": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "integer"
                }
            }
        },
        "domain.GetAllOrders": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "cost": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "customerID": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "orderDishes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.OrderDish"
                    }
                },
                "statusID": {
                    "type": "integer"
                }
            }
        },
        "domain.GetOrderByID": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "cost": {
                    "type": "number"
                },
                "courierID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "customerID": {
                    "type": "integer"
                },
                "factTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "orderDishes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.OrderDish"
                    }
                },
                "paymentID": {
                    "type": "string"
                },
                "requiredTime": {
                    "type": "string"
                },
                "statusID": {
                    "type": "integer"
                }
            }
        },
        "domain.Order": {
            "type": "object",
            "required": [
                "address",
                "customerID",
                "dishes",
                "requiredTime",
                "restaurantID"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "customerID": {
                    "type": "integer"
                },
                "dishes": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/domain.OrderDish"
                    }
                },
                "requiredTime": {
                    "type": "string"
                },
                "restaurantID": {
                    "type": "string"
                }
            }
        },
        "domain.OrderDish": {
            "type": "object",
            "required": [
                "amount",
                "dishID"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "dishID": {
                    "type": "string"
                }
            }
        },
        "handler.customerResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/domain.CustomerInfo"
                }
            }
        },
        "handler.dataResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
        "handler.orderResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/domain.GetOrderByID"
                }
            }
        },
        "handler.ordersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.GetAllOrders"
                    }
                }
            }
        },
        "handler.response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}